Fix table panel
Table panel uses Object.keys which is not supported by lazy messages
since they do not have enumerable own properties. Fix by using toJSON()
on the lazy message which returns a plain object.

Update lazy message cache properties to be private so they don't show
up in Object.keys and create confusion as to why these weird properties
are present.

Fixes #931
