+++ .eslintrc.yaml
@@ -301,9 +301,6 @@ overrides:
       - "app/panels/Tab/ToolbarTab.tsx"
       - "app/panels/Tab/index.stories.tsx"
       - "app/panels/Tab/index.tsx"
-      - "app/panels/Table/index.stories.tsx"
-      - "app/panels/Table/index.tsx"
-      - "app/panels/Table/types.ts"
       - "app/panels/ThreeDimensionalViz/CameraInfo/index.stories.tsx"
       - "app/panels/ThreeDimensionalViz/CameraInfo/index.tsx"
       - "app/panels/ThreeDimensionalViz/DebugStats.tsx"
