+++ app/panels/Table/index.stories.tsx
@@ -16,7 +16,7 @@ import { storiesOf } from "@storybook/react";
 import Table from "@foxglove-studio/app/panels/Table";
 import PanelSetup from "@foxglove-studio/app/stories/PanelSetup";
 
-const makeArrayData = (length = 50, nestArray = true): any => {
+const makeArrayData = (length = 50, nestArray = true): unknown => {
   return new Array(length).fill(0).map((_, i) => {
     return {
       val: i,
@@ -79,7 +79,9 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=expand-row-0]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=expand-row-0]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
@@ -93,7 +95,9 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=expand-cell-obj-0]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=expand-cell-obj-0]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
@@ -107,7 +111,9 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=expand-cell-arr-0]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=expand-cell-arr-0]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
@@ -121,8 +127,12 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=expand-row-0]")[0] as any).click();
-            (document.querySelectorAll("[data-test=expand-cell-arr-obj-0]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=expand-row-0]",
+            )[0] as HTMLTableCellElement).click();
+            (document.querySelectorAll(
+              "[data-test=expand-cell-arr-obj-0]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
@@ -136,8 +146,12 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=expand-row-0]")[0] as any).click();
-            (document.querySelectorAll("[data-test=expand-row-1]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=expand-row-0]",
+            )[0] as HTMLTableCellElement).click();
+            (document.querySelectorAll(
+              "[data-test=expand-row-1]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
@@ -158,8 +172,12 @@ storiesOf("panels/Table/index", module)
         fixture={fixture}
         onMount={() => {
           setImmediate(() => {
-            (document.querySelectorAll("[data-test=column-header-val]")[0] as any).click();
-            (document.querySelectorAll("[data-test=column-header-val]")[0] as any).click();
+            (document.querySelectorAll(
+              "[data-test=column-header-val]",
+            )[0] as HTMLTableCellElement).click();
+            (document.querySelectorAll(
+              "[data-test=column-header-val]",
+            )[0] as HTMLTableCellElement).click();
           });
         }}
       >
